Move component ConfirmedCasesDetailsCard from 'components/cards' to 'components/index/CardsMonitoring'
