+++ components/index/CardsMonitoring/ConfirmedCasesDetailsCard.vue
